% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkG12.R
\name{mkG12}
\alias{mkG12}
\title{Makes generalized additive models for causes of death that were aggregated into 12 groups}
\usage{
mkG12(
  seerHome = "~/data/CMLepi",
  inFile = "seerMrt.RData",
  outFile = "G12.RData"
)
}
\arguments{
\item{seerHome}{folder name.}

\item{inFile}{input file.}

\item{outFile}{output file.}
}
\value{
Nothing is returned. Run for the side effect of creating a list of gam fits in the output file.
}
\description{
Gives each covid year a separate fit across ages.
}
