% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkG13.R
\name{mkG13}
\alias{mkG13}
\title{Makes Generalized Additive Model for 14 cause of death groups}
\usage{
mkG13(
  seerHome = "~/data/CMLepi",
  inFile = "seerMrt.RData",
  outFile = "G13.RData"
)
}
\arguments{
\item{seerHome}{folder name.}

\item{inFile}{input file.}

\item{outFile}{output file.}
}
\value{
Nothing is returned. Run for the side effect of creating a list of 12 gam fits in the output file.
}
\description{
Makes Generalized Additive Model for 14 cause of death groups
}
