% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer2r.R
\name{seer2r}
\alias{seer2r}
\title{Convert db.txt db.dic files generated in a SEER*stat Case Listing session into an R tibble}
\usage{
seer2r(db, path = "~/data/CMLepi")
}
\arguments{
\item{db}{database name.}

\item{path}{folder name.}
}
\value{
A tibble.
}
\description{
Convert db.txt db.dic files generated in a SEER*stat Case Listing session into an R tibble
}
